pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iot-secure-framework"
version = "1.0.0"
description = "Two-Phase IoT Security Framework: Network Profiling + Self-Labeled Federated Learning IDS"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "IoT Security Team", email = "team@iot-security.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "scapy>=2.5.0",
    "dpkt>=1.9.8",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "joblib>=1.3.0",
    "pyarrow>=12.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]

[project.scripts]
iot-phase1 = "src.phase1_profiling.cli:main"
iot-phase2 = "src.phase2_ids.cli:main"
iot-service = "apps.service:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "apps*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
